Fix duplicate data variable declaration and foreign key constraint type mismatch

+++ database/migrations/2025_12_21_032000_add_user_targeting_to_alerts.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('alert_user', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('alert_id');
-            $table->unsignedBigInteger('user_id');
+            $table->unsignedInteger('user_id'); // Changed from unsignedBigInteger to match users.id type
             $table->timestamps();
 
             $table->foreign('alert_id')->references('id')->on('alerts')->onDelete('cascade');
